==> 0_SP500StockList.ipynb <==
import pandas as pd

filename='sp500_list.csv'
# URL to the S&P 500 Wikipedia page
url = 'https://en.wikipedia.org/wiki/List_of_S%26P_500_companies'

# Read the table from the Wikipedia page
tables = pd.read_html(url)
sp500_table = tables[0]  # The first table contains the S&P 500 list

sp500_table = sp500_table[['Symbol', 'Security', 'GICS Sub-Industry']]
sp500_table.columns = ['Ticker', 'Company Name', 'Industry']

# Save the table to a CSV file
sp500_table.to_csv(filename, index=False)
print(f"S&P 500 stock list saved to {filename}")
